requirements.txt
flask
python-dotenv
telethon
pymongo
motor